Library

Contents

Library

Contents

Index

Public

Index

DisjointCliqueCover.get_EECC

DisjointCliqueCover.limited_maximal_cliques

Public

DisjointCliqueCover.get_EECC — Method

get_EECC(G::SimpleGraph, m₀::Int64)

Calculate the edge-disjoint edge clique cover (EECC) of a graph G considering cliques of order up to m₀, according to the heuristic proposed in the paper.

Arguments

G::SimpleGraph: a simple graph of the LightGraphs.SimpleGraph type

m₀::Int64: an integer representing the maximum order of the cliques to consider

Return

A vector containing the EECC

DisjointCliqueCover.limited_maximal_cliques — Method

limited_maximal_cliques(G::SimpleGraph, m₀::Int64)

Calculate the maximal cliques of a graph G up to order m₀. Any clique of order m>m₀ is decomposed into its Base.binomial(m, m₀) sub-cliques of order m₀.

Arguments

G::SimpleGraph: a simple graph of the LightGraphs.SimpleGraph type

m₀::Int64: an integer representing the maximum order of the cliques to consider

Return

A vector containing the maximal cliques up to order m₀